+++ src/components/revoke-block.tsx
@@ -1,9 +1,9 @@
 import React, { FunctionComponent, useState } from "react";
-import { TextInput } from "./common/TextInput";
-import { SecondaryButton, PrimaryButton } from "./common/Button";
-import { ConfirmationModalDialog } from "./common/ModalDialog";
-import { Spinner } from "./common/Spinner";
-import { Logger } from "./common/Logger";
+import { TextInput } from "./common/text-input";
+import { SecondaryButton, PrimaryButton } from "./common/button";
+import { ConfirmationModalDialog } from "./common/modal-dialog";
+import { Spinner } from "./common/spinner";
+import { Logger } from "./common/logger";
 import { isValidHash, getEtherscanAddress } from "./util/common";
 import { revokeCertificateHash as revoke } from "./util/revoke";
 import { getWalletNetwork } from "./util/wallet";
